requirements.txt
openai>=1.0.0
redis>=4.0.0
python-dotenv>=0.19.0
rasa-sdk>=3.1.0 